+++ src/DataLoad.h
@@ -139,7 +139,7 @@ class DataLoad : public DataGeneric
     void change_p(int load_id, real_type new_p, bool & need_reset);
     void change_q(int load_id, real_type new_q, bool & need_reset);
 
-    virtual void fillSbus(CplxVect & Sbus, const std::vector<int> & id_grid_to_solver) const;
+    virtual void fillSbus(CplxVect & Sbus, const std::vector<int> & id_grid_to_solver, bool ac) const;
 
     void compute_results(const Eigen::Ref<const RealVect> & Va,
                          const Eigen::Ref<const RealVect> & Vm,