+++ src/DataTrafo.h
@@ -180,6 +180,11 @@ class DataTrafo : public DataGeneric
                           bool ac,
                           const std::vector<int> & id_grid_to_solver,
                           real_type sn_mva) const;
+    virtual void fillBp_Bpp(std::vector<Eigen::Triplet<real_type> > & Bp,
+                            std::vector<Eigen::Triplet<real_type> > & Bpp,
+                            const std::vector<int> & id_grid_to_solver,
+                            real_type sn_mva,
+                            FDPFMethod xb_or_bx) const;
     virtual void hack_Sbus_for_dc_phase_shifter(CplxVect & Sbus, bool ac, const std::vector<int> & id_grid_to_solver);  // needed for dc mode
 
     void compute_results(const Eigen::Ref<const RealVect> & Va,